Improved UI for removing relations.

+++ app/src/main/java/com/individual/vozenredmk/FavoritesRelationAdapter.java
@@ -4,23 +4,34 @@ import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
 import android.widget.CompoundButton;
 import android.widget.Toast;
+import android.widget.ToggleButton;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
+import java.util.Objects;
 
 public class FavoritesRelationAdapter extends RecyclerView.Adapter<RelacijaViewHolder> {
     Context contextFavorites;
     ArrayList<Relation> relationFavorites;
     RecyclerView recyclerViewFavorites;
+    ToggleButton heartButton;
     Fragment f;
+    DBHelper dbHelper;
+
 
     public FavoritesRelationAdapter(Context contextFavorites, ArrayList<Relation> relationFavorites) {
+
         this.contextFavorites = contextFavorites;
         this.relationFavorites = relationFavorites;
 
@@ -31,8 +42,14 @@ public class FavoritesRelationAdapter extends RecyclerView.Adapter<RelacijaViewH
     @NonNull
     @Override
     public RelacijaViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+        dbHelper = new DBHelper(contextFavorites);
         View view = LayoutInflater.from(contextFavorites).inflate(R.layout.fragment_favorites, parent, false);
-        recyclerViewFavorites = view.findViewById(R.id.recyclerViewFavorites);
+//        recyclerViewFavorites = view.findViewById(R.id.recyclerViewFavorites);
+//        recyclerViewFavorites.setHasFixedSize(true);
+//        recyclerViewFavorites.setLayoutManager(new LinearLayoutManager(contextFavorites));
+//        relationFavorites = (ArrayList<Relation>) dbHelper.getEveryone();
+//
+
         return new RelacijaViewHolder(LayoutInflater.from(contextFavorites).inflate(R.layout.relacii_cardview, parent, false));
 
     }
@@ -40,7 +57,7 @@ public class FavoritesRelationAdapter extends RecyclerView.Adapter<RelacijaViewH
 
 
     public void onBindViewHolder(@NonNull final RelacijaViewHolder holder, final int position) {
-        final DBHelper dbHelper = new DBHelper(contextFavorites);
+        dbHelper = new DBHelper(contextFavorites);
         holder.relacija.setText(relationFavorites.get(position).getRelacija());
         holder.vremeikompanija.setText(relationFavorites.get(position).getVremeIKompanija());
         holder.cena.setText(relationFavorites.get(position).getCena());
@@ -53,13 +70,20 @@ public class FavoritesRelationAdapter extends RecyclerView.Adapter<RelacijaViewH
         holder.toggleButton.setChecked(true);
         holder.toggleButton.setBackgroundDrawable(ContextCompat.getDrawable(contextFavorites.getApplicationContext(), R.drawable.ic_baseline_favorite_24));
         holder.toggleButton.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+
+            Relation foundIt = dbHelper.getSingleRow(relationFavorites.get(holder.getAdapterPosition()).getId());
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                Relation foundIt = dbHelper.getSingleRow(relationFavorites.get(position).getId());
-                dbHelper.deleteOne(foundIt);
-                relationFavorites = (ArrayList<Relation>) dbHelper.getEveryone();
-                recyclerViewFavorites.setAdapter(new FavoritesRelationAdapter(contextFavorites, relationFavorites));
-                recyclerViewFavorites.getAdapter();
+                if(isChecked) {
+                    holder.toggleButton.setBackgroundDrawable(ContextCompat.getDrawable(contextFavorites.getApplicationContext(), R.drawable.ic_baseline_favorite_24));
+                    dbHelper.addOne(foundIt);
+
+                }
+                else {
+                    holder.toggleButton.setBackgroundDrawable(ContextCompat.getDrawable(contextFavorites.getApplicationContext(), R.drawable.ic_baseline_favorite_border_24));
+                    dbHelper.deleteOne(foundIt);
+                }
+
             }
         });
     }
